requirements.txt
fastapi
uvicorn
motor
boto3
streamlit
python-dotenv
pydantic
pydantic-settings
requests
apscheduler
